working on up and downvote

+++ src/main/java/com/vervloet/msgservices/service/MessageService.java
@@ -6,6 +6,7 @@ import com.vervloet.msgservices.exception.ResourceNotFoundException;
 import com.vervloet.msgservices.model.User;
 import com.vervloet.msgservices.persistence.MessageRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -41,11 +42,64 @@ public class MessageService {
                 .orElseThrow(() -> new ResourceNotFoundException("message", "id", messageId));
     }
 
+    public ResponseEntity<?> upvoteMessage(Long messageId) {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        CustomUserDetails customUserDetails = ((CustomUserDetails)principal);
+
+        Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new ResourceNotFoundException("message", "id", messageId));
+
+        if(message.checkVoted(customUserDetails.getUsername())){
+
+            return new ResponseEntity<>("Ja votado", HttpStatus.FORBIDDEN);
+
+        } else {
+
+            message.setVotes(message.getVotes() + 1L);
+            message.addVoted(customUserDetails.getUsername());
+
+            return ResponseEntity.ok().build();
+        }
+    }
+
+    public ResponseEntity<?> downvoteMessage(Long messageId) {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        CustomUserDetails customUserDetails = ((CustomUserDetails)principal);
+
+        Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new ResourceNotFoundException("message", "id", messageId));
+
+        if(message.checkVoted(customUserDetails.getUsername())){
+
+            return new ResponseEntity<>("Ja votado", HttpStatus.FORBIDDEN);
+
+        } else {
+
+            message.setVotes(message.getVotes() - 1L);
+            message.addVoted(customUserDetails.getUsername());
+
+            return ResponseEntity.ok().build();
+        }
+    }
+
     public ResponseEntity<?> deleteMessage(Long messageId) {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        CustomUserDetails customUserDetails = ((CustomUserDetails)principal);
+
         Message message = messageRepository.findById(messageId)
                 .orElseThrow(() -> new ResourceNotFoundException("Message", "id", messageId));
 
-        messageRepository.delete(message);
+        if( message.getUsername().equals(customUserDetails.getUsername())){
+            messageRepository.delete(message);
+        } else {
+            return new ResponseEntity<>("Nao Autorizado", HttpStatus.UNAUTHORIZED);
+        }
 
         return ResponseEntity.ok().build();
     }
